+++ CHANGES.md
@@ -8,6 +8,7 @@ Copyright (c) 2014-2018, CKSource - Frederico Knabben. All rights reserved.
 ## Version 1.1.1
 
 * [#233](https://github.com/cksource/ckeditor-plugin-a11ychecker/issues/233): Fixed: Balloon classes are localized, causing issues of a different testability look the same.
+* [#247](https://github.com/cksource/ckeditor-plugin-a11ychecker/issues/247): Introduced `process` event in the `Engine` type, allowing for adding custom issue types.
 
 ## Version 1.1.0